net/netvsc: fix ethdev get_supported_ptypes
When doing rte_eth_dev_get_supported_ptypes on netvsc device
the values reported are incorrect if VF is not present.

If no VF is present the device uses rte_net_get_ptype()
which can return a wide range of values. Use the same table
as TAP device in that case.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>