net/bnxt: don't fail rte_pmd_bnxt_set_all_queues_drop_en if no VF's
The function should return success if called with no active VF's.
It is OK to call this function if the desired outcome is to change
the flag for PF's even if there are no VF's.

Original code may have worked by accidental uninitialized variable;
then this bug was introduced by change to fix compilation warning.

Fixes: 4a671fdea9f4 ("net/bnxt: fix compilation with -Og")
Cc: olivier.matz@6wind.com
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>