net/mlx5:fix storing the synced MAC to internal table
As the internal MAC table is divided into Unicast and Multicast address
section, we should check the type of synced MAC address before storing
it to the internal table. Currently the check is not done, and the
synced MAC of 33:33:00:00:00:01 gets stored in the unicast section
(mostly index 1) which causes all subsequent mlx5_set_mc_addr_list()
to fail with error -EADDRINUSE, as the mac_list contains the MAC
33:33:00:00:00:01. This denies adding of any new multicast address to
the internal list and also fails to add the MAC address to the device
in case of SR-IOV VF case.

Signed-off-by: Souvik Dey <sodey@rbbn.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>