net/ixgbe: fix fdirctrl register setting
The function ixgbe_fdir_set_flexbytes_offset is used when create FDir
rule for flexbytes. It set a register: FDIRCTRL.FLEX_OFFSET, which
cause that even if the FDir flexbytes rule is destroyed, the rule still
direct the packet and transfer it to the wrong place. It is because
Setting FDIRCTRL shall only be permitted on Flow Director
initialization flow or Clearing the Flow Director table, otherwise
unexpected happens. In order to evade the limit, add code to make
setting FDIRCTRL work without unexpected effects.

Fixes: f35fec63dde1 ("net/ixgbe: enable flex bytes for generic flow API")
Cc: qi.z.zhang@intel.com

Signed-off-by: YU DAPENG <dapengx.yu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>