net/ixgbe: fix fdirctrl register setting
The function ixgbe_fdir_set_flexbytes_offset is used when create FDir
rule for flexbytes. It set a register: FDIRCTRL.FLEX_OFFSET, which
cause that even if the FDir flexbytes rule is destroyed, the rule still
direct the packet and transfer it to the wrong place. It is because
setting FDIRCTRL shall only be permitted on Flow Director
initialization flow or clearing the Flow Director table according to
intel datasheet, otherwise unexpected happens. In order to evade the
limit, add code to set FDIRCMD.CLEARHT to 1b and then clear it back to
0b to make the setting act like the Flow Director initialization flow
or clearing the Flow Director table.

Fixes: f35fec63dde1 ("net/ixgbe: enable flex bytes for generic flow API")
Cc: stable@dpdk.org

Signed-off-by: YU DAPENG <dapengx.yu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>