power: fix make build for power apps
The guest channel message definitions and commands in guest_channel.h file
are needed by applications and need to be made public.

This patch is needed because make is broken for the power_mgmt apps, as
the channel commands API is private. This previously built within DPDK,
but now with all the meson/ninja changes, making these apps externally
no longer works. To fix, we need to move the header file with the API
definitions for the channel commands public, and rename the functionss
accordingly.

The main change is to rename channel_commands.h to
rte_power_guest_channel.h so that it gets picked up by the installer and
copied to /usr/local/include. Other changes include renaming #defines to
have RTE_ at the beginning instead of CPU_. Finally we refactor the code
to work with those changes.

Fixes: 210c383e247b ("power: packet format for vm power management")
Fixes: cd0d5547e873 ("power: vm communication channels in guest")
Cc: stable@dpdk.org

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: David Hunt <david.hunt@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>